backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-aiplatform==1.60.0
google-cloud-firestore==2.16.0
pydantic==2.7.1
python-dotenv==1.0.1
